staging/pkg/mod/golang.org/x/net@v0.0.0-20201110031124-69a78807bb2b/html/testdata/go/issue_30961_error_nested_unknown_tag_types.dat

#data
<html><head></head><body><tag1><tag2 /><p></p></tag1><div></div></body></html>
#errors
#document
| <html>
|   <head>
|   <body>
|     <tag1>
|       <tag2>
|         <p>
|     <div>
